refactor: consolidate .d.ts files into a single index.d.ts
Merges all separate type declaration files into lib/index.d.ts to
simplify maintenance. Also adds an explicit "types" field to package.json
for unambiguous TypeScript resolution.
