--- lib/createChoices.d.ts
@@ -1,2 +1,0 @@
-import { Choice, Summary } from './types'
-export declare function createChoices({ packageSummary }: { packageSummary: Summary[] }): Choice[]
