--- lib/createPackageSummary.d.ts
@@ -1,2 +1,0 @@
-import { Summary } from './types'
-export declare function createPackageSummary(options: { cwd?: string; pkg: any }): Promise<Summary[]>
