--- lib/interactiveUpdate.d.ts
@@ -1,1 +1,0 @@
-export declare function interactiveUpdate(options: { cwd?: string; update?: boolean }): Promise<void>
